+++ backend/src/main/java/ssafy/haruman/domain/deposit/dto/response/DepositDetailResponseDto.java
@@ -4,8 +4,6 @@ package ssafy.haruman.domain.deposit.dto.response;
 import lombok.*;
 import ssafy.haruman.domain.deposit.entity.Deposit;
 
-import java.math.BigDecimal;
-
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
@@ -15,7 +13,7 @@ public class DepositDetailResponseDto {
     private String bank;
     private String name;
     private String description;
-    private BigDecimal interestRate;
+    private float interestRate;
 
 
     public static DepositDetailResponseDto from(Deposit deposit) {
